raidemulator: Simplify worker loading
