+++ types/worker.d.ts
@@ -1,7 +1,20 @@
-declare module '*/NetworkLogConverter.worker.ts' {
-  export class NetworkLogConverterWorker extends Worker {
-    constructor();
-  }
+import Encounter from '../ui/raidboss/emulator/data/Encounter';
 
-  export default NetworkLogConverterWorker;
-}
+export type DoneMessage = {
+  type: 'done';
+};
+
+export type EncounterMessage = {
+  type: 'encounter';
+  encounter: Encounter;
+  name: string;
+};
+
+export type ProgressMessage = {
+  type: 'progress';
+  lines: number;
+  bytes: number;
+  totalBytes: number;
+};
+
+export type ConverterWorkerMessage = DoneMessage | EncounterMessage | ProgressMessage;
